@@ -2,100 +2,157 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.9
-
 import 'dart:async';
+import 'dart:html' as html;
 
 import 'package:integration_test/integration_test.dart';
 import 'package:google_maps/google_maps.dart' as gmaps;
 import 'package:google_maps_flutter_web/google_maps_flutter_web.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/mockito.dart';
-
-class _MockMarker extends Mock implements gmaps.Marker {
-  final onClickController = StreamController<gmaps.MouseEvent>();
-  final onDragEndController = StreamController<gmaps.MouseEvent>();
-
-  @override
-  Stream<gmaps.MouseEvent> get onClick => onClickController.stream;
-
-  @override
-  Stream<gmaps.MouseEvent> get onDragend => onDragEndController.stream;
-}
-
-class _MockMouseEvent extends Mock implements gmaps.MouseEvent {}
-
-class _MockInfoWindow extends Mock implements gmaps.InfoWindow {}
 
 /// Test Markers
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  bool called = false;
+  // Since onTap/DragEnd events happen asynchronously, we need to store when the event
+  // is fired. We use a completer so the test can wait for the future to be completed.
+  late Completer<bool> _methodCalledCompleter;
+
+  /// This is the future value of the [_methodCalledCompleter]. Reinitialized
+  /// in the [setUp] method, and completed (as `true`) by [onTap] and [onDragEnd]
+  /// when those methods are called from the MarkerController.
+  late Future<bool> methodCalled;
+
   void onTap() {
-    called = true;
+    _methodCalledCompleter.complete(true);
   }
 
   void onDragEnd(gmaps.LatLng _) {
-    called = true;
+    _methodCalledCompleter.complete(true);
   }
 
   setUp(() {
-    called = false;
+    _methodCalledCompleter = Completer();
+    methodCalled = _methodCalledCompleter.future;
   });
 
   group('MarkerController', () {
-    _MockMarker marker;
+    late gmaps.Marker marker;
 
     setUp(() {
-      marker = _MockMarker();
+      marker = gmaps.Marker();
     });
 
     testWidgets('onTap gets called', (WidgetTester tester) async {
       MarkerController(marker: marker, onTap: onTap);
-      // Simulate a click
-      await marker.onClickController.add(null);
-      expect(called, isTrue);
+
+      // Trigger a click event...
+      gmaps.Event.trigger(marker, 'click', [gmaps.MapMouseEvent()]);
+
+      // The event handling is now truly async. Wait for it...
+      expect(await methodCalled, isTrue);
     });
 
     testWidgets('onDragEnd gets called', (WidgetTester tester) async {
-      when(marker.draggable).thenReturn(true);
       MarkerController(marker: marker, onDragEnd: onDragEnd);
-      // Simulate a drag end
-      await marker.onDragEndController.add(_MockMouseEvent());
-      expect(called, isTrue);
+
+      // Trigger a drag end event...
+      gmaps.Event.trigger(marker, 'dragend',
+          [gmaps.MapMouseEvent()..latLng = gmaps.LatLng(0, 0)]);
+
+      expect(await methodCalled, isTrue);
     });
 
     testWidgets('update', (WidgetTester tester) async {
       final controller = MarkerController(marker: marker);
-      final options = gmaps.MarkerOptions()..draggable = false;
+      final options = gmaps.MarkerOptions()..draggable = true;
+
+      expect(marker.draggable, isNull);
+
       controller.update(options);
-      verify(marker.options = options);
+
+      expect(marker.draggable, isTrue);
     });
 
     testWidgets('infoWindow null, showInfoWindow.',
         (WidgetTester tester) async {
       final controller = MarkerController(marker: marker);
+
       controller.showInfoWindow();
+
       expect(controller.infoWindowShown, isFalse);
     });
 
     testWidgets('showInfoWindow', (WidgetTester tester) async {
-      final infoWindow = _MockInfoWindow();
+      final infoWindow = gmaps.InfoWindow();
+      final map = gmaps.GMap(html.DivElement());
+      marker.set('map', map);
       final controller =
           MarkerController(marker: marker, infoWindow: infoWindow);
+
       controller.showInfoWindow();
-      verify(infoWindow.open(any, any)).called(1);
+
+      expect(infoWindow.get('map'), map);
       expect(controller.infoWindowShown, isTrue);
     });
 
     testWidgets('hideInfoWindow', (WidgetTester tester) async {
-      final infoWindow = _MockInfoWindow();
+      final infoWindow = gmaps.InfoWindow();
+      final map = gmaps.GMap(html.DivElement());
+      marker.set('map', map);
       final controller =
           MarkerController(marker: marker, infoWindow: infoWindow);
+
       controller.hideInfoWindow();
-      verify(infoWindow.close()).called(1);
+
+      expect(infoWindow.get('map'), isNull);
       expect(controller.infoWindowShown, isFalse);
+    });
+
+    group('remove', () {
+      late MarkerController controller;
+
+      setUp(() {
+        final infoWindow = gmaps.InfoWindow();
+        final map = gmaps.GMap(html.DivElement());
+        marker.set('map', map);
+        controller = MarkerController(marker: marker, infoWindow: infoWindow);
+      });
+
+      testWidgets('drops gmaps instance', (WidgetTester tester) async {
+        controller.remove();
+
+        expect(controller.marker, isNull);
+      });
+
+      testWidgets('cannot call update after remove',
+          (WidgetTester tester) async {
+        final options = gmaps.MarkerOptions()..draggable = true;
+
+        controller.remove();
+
+        expect(() {
+          controller.update(options);
+        }, throwsAssertionError);
+      });
+
+      testWidgets('cannot call showInfoWindow after remove',
+          (WidgetTester tester) async {
+        controller.remove();
+
+        expect(() {
+          controller.showInfoWindow();
+        }, throwsAssertionError);
+      });
+
+      testWidgets('cannot call hideInfoWindow after remove',
+          (WidgetTester tester) async {
+        controller.remove();
+
+        expect(() {
+          controller.hideInfoWindow();
+        }, throwsAssertionError);
+      });
     });
   });
 }
