@@ -27,11 +27,11 @@ class GoogleMapController {
   String _getViewType(int mapId) => 'plugins.flutter.io/google_maps_$mapId';
 
   // The Flutter widget that contains the rendered Map.
-  HtmlElementView _widget;
-  HtmlElement _div;
+  HtmlElementView? _widget;
+  late HtmlElement _div;
 
   /// The Flutter widget that will contain the rendered Map. Used for caching.
-  HtmlElementView get widget {
+  Widget? get widget {
     if (_widget == null && !_streamController.isClosed) {
       _widget = HtmlElementView(
         viewType: _getViewType(_mapId),
@@ -41,14 +41,14 @@ class GoogleMapController {
   }
 
   // The currently-enabled traffic layer.
-  gmaps.TrafficLayer _trafficLayer;
+  gmaps.TrafficLayer? _trafficLayer;
 
   /// A getter for the current traffic layer. Only for tests.
   @visibleForTesting
-  gmaps.TrafficLayer get trafficLayer => _trafficLayer;
+  gmaps.TrafficLayer? get trafficLayer => _trafficLayer;
 
   // The underlying GMap instance. This is the interface with the JS SDK.
-  gmaps.GMap _googleMap;
+  gmaps.GMap? _googleMap;
 
   // The StreamController used by this controller and the geometry ones.
   final StreamController<MapEvent> _streamController;
@@ -57,10 +57,10 @@ class GoogleMapController {
   Stream<MapEvent> get events => _streamController.stream;
 
   // Geometry controllers, for different features of the map.
-  CirclesController _circlesController;
-  PolygonsController _polygonsController;
-  PolylinesController _polylinesController;
-  MarkersController _markersController;
+  CirclesController? _circlesController;
+  PolygonsController? _polygonsController;
+  PolylinesController? _polylinesController;
+  MarkersController? _markersController;
   // Keeps track if _attachGeometryControllers has been called or not.
   bool _controllersBoundToMap = false;
 
@@ -69,9 +69,9 @@ class GoogleMapController {
 
   /// Initializes the GMap, and the sub-controllers related to it. Wires events.
   GoogleMapController({
-    @required int mapId,
-    @required StreamController<MapEvent> streamController,
-    @required CameraPosition initialCameraPosition,
+    required int mapId,
+    required StreamController<MapEvent> streamController,
+    required CameraPosition initialCameraPosition,
     Set<Marker> markers = const <Marker>{},
     Set<Polygon> polygons = const <Polygon>{},
     Set<Polyline> polylines = const <Polyline>{},
@@ -107,11 +107,11 @@ class GoogleMapController {
   /// Overrides certain properties to install mocks defined during testing.
   @visibleForTesting
   void debugSetOverrides({
-    DebugCreateMapFunction createMap,
-    MarkersController markers,
-    CirclesController circles,
-    PolygonsController polygons,
-    PolylinesController polylines,
+    DebugCreateMapFunction? createMap,
+    MarkersController? markers,
+    CirclesController? circles,
+    PolygonsController? polygons,
+    PolylinesController? polylines,
   }) {
     _overrideCreateMap = createMap;
     _markersController = markers ?? _markersController;
@@ -120,11 +120,11 @@ class GoogleMapController {
     _polylinesController = polylines ?? _polylinesController;
   }
 
-  DebugCreateMapFunction _overrideCreateMap;
+  DebugCreateMapFunction? _overrideCreateMap;
 
   gmaps.GMap _createMap(HtmlElement div, gmaps.MapOptions options) {
     if (_overrideCreateMap != null) {
-      return _overrideCreateMap(div, options);
+      return _overrideCreateMap!(div, options);
     }
     return gmaps.GMap(div, options);
   }
@@ -142,10 +142,11 @@ class GoogleMapController {
     options = _applyInitialPosition(_initialCameraPosition, options);
 
     // Create the map...
-    _googleMap = _createMap(_div, options);
+    final map = _createMap(_div, options);
+    _googleMap = map;
 
-    _attachMapEvents(_googleMap);
-    _attachGeometryControllers(_googleMap);
+    _attachMapEvents(map);
+    _attachGeometryControllers(map);
 
     _renderInitialGeometry(
       markers: _markers,
@@ -154,19 +155,21 @@ class GoogleMapController {
       polylines: _polylines,
     );
 
-    _setTrafficLayer(_googleMap, _isTrafficLayerEnabled(_rawMapOptions));
+    _setTrafficLayer(map, _isTrafficLayerEnabled(_rawMapOptions));
   }
 
   // Funnels map gmap events into the plugin's stream controller.
   void _attachMapEvents(gmaps.GMap map) {
     map.onClick.listen((event) {
+      assert(event.latLng != null);
       _streamController.add(
-        MapTapEvent(_mapId, _gmLatLngToLatLng(event.latLng)),
+        MapTapEvent(_mapId, _gmLatLngToLatLng(event.latLng!)),
       );
     });
     map.onRightclick.listen((event) {
+      assert(event.latLng != null);
       _streamController.add(
-        MapLongPressEvent(_mapId, _gmLatLngToLatLng(event.latLng)),
+        MapLongPressEvent(_mapId, _gmLatLngToLatLng(event.latLng!)),
       );
     });
     map.onBoundsChanged.listen((event) {
@@ -188,28 +191,47 @@ class GoogleMapController {
   void _attachGeometryControllers(gmaps.GMap map) {
     // Now we can add the initial geometry.
     // And bind the (ready) map instance to the other geometry controllers.
-    _circlesController.bindToMap(_mapId, map);
-    _polygonsController.bindToMap(_mapId, map);
-    _polylinesController.bindToMap(_mapId, map);
-    _markersController.bindToMap(_mapId, map);
+    //
+    // These controllers are either created in the constructor of this class, or
+    // overriden (for testing) by the [debugSetOverrides] method. They can't be
+    // null.
+    assert(_circlesController != null,
+        'Cannot attach a map to a null CirclesController instance.');
+    assert(_polygonsController != null,
+        'Cannot attach a map to a null PolygonsController instance.');
+    assert(_polylinesController != null,
+        'Cannot attach a map to a null PolylinesController instance.');
+    assert(_markersController != null,
+        'Cannot attach a map to a null MarkersController instance.');
+
+    _circlesController!.bindToMap(_mapId, map);
+    _polygonsController!.bindToMap(_mapId, map);
+    _polylinesController!.bindToMap(_mapId, map);
+    _markersController!.bindToMap(_mapId, map);
+
     _controllersBoundToMap = true;
   }
 
   // Renders the initial sets of geometry.
   void _renderInitialGeometry({
-    Set<Marker> markers,
-    Set<Circle> circles,
-    Set<Polygon> polygons,
-    Set<Polyline> polylines,
+    Set<Marker> markers = const {},
+    Set<Circle> circles = const {},
+    Set<Polygon> polygons = const {},
+    Set<Polyline> polylines = const {},
   }) {
     assert(
         _controllersBoundToMap,
         'Geometry controllers must be bound to a map before any geometry can ' +
             'be added to them. Ensure _attachGeometryControllers is called first.');
-    _markersController.addMarkers(markers);
-    _circlesController.addCircles(circles);
-    _polygonsController.addPolygons(polygons);
-    _polylinesController.addPolylines(polylines);
+
+    // The above assert will only succeed if the controllers have been bound to a map
+    // in the [_attachGeometryControllers] method, which ensures that all these
+    // controllers below are *not* null.
+
+    _markersController!.addMarkers(markers);
+    _circlesController!.addCircles(circles);
+    _polygonsController!.addPolygons(polygons);
+    _polylinesController!.addPolylines(polylines);
   }
 
   // Merges new options coming from the plugin into the _rawMapOptions map.
@@ -227,10 +249,12 @@ class GoogleMapController {
   ///
   /// This method converts the map into the proper [gmaps.MapOptions]
   void updateRawOptions(Map<String, dynamic> optionsUpdate) {
+    assert(_googleMap != null, 'Cannot update options on a null map.');
+
     final newOptions = _mergeRawOptions(optionsUpdate);
 
     _setOptions(_rawOptionsToGmapsOptions(newOptions));
-    _setTrafficLayer(_googleMap, _isTrafficLayerEnabled(newOptions));
+    _setTrafficLayer(_googleMap!, _isTrafficLayerEnabled(newOptions));
   }
 
   // Sets new [gmaps.MapOptions] on the wrapped map.
@@ -241,11 +265,10 @@ class GoogleMapController {
   // Attaches/detaches a Traffic Layer on the passed `map` if `attach` is true/false.
   void _setTrafficLayer(gmaps.GMap map, bool attach) {
     if (attach && _trafficLayer == null) {
-      _trafficLayer = gmaps.TrafficLayer();
-      _trafficLayer.set('map', map);
+      _trafficLayer = gmaps.TrafficLayer()..set('map', map);
     }
     if (!attach && _trafficLayer != null) {
-      _trafficLayer.set('map', null);
+      _trafficLayer!.set('map', null);
       _trafficLayer = null;
     }
   }
@@ -255,35 +278,61 @@ class GoogleMapController {
 
   /// Returns the [LatLngBounds] of the current viewport.
   Future<LatLngBounds> getVisibleRegion() async {
-    return _gmLatLngBoundsTolatLngBounds(await _googleMap.bounds);
+    assert(_googleMap != null, 'Cannot get the visible region of a null map.');
+
+    return _gmLatLngBoundsTolatLngBounds(
+      await _googleMap!.bounds ?? _nullGmapsLatLngBounds,
+    );
   }
 
   /// Returns the [ScreenCoordinate] for a given viewport [LatLng].
   Future<ScreenCoordinate> getScreenCoordinate(LatLng latLng) async {
+    assert(_googleMap != null,
+        'Cannot get the screen coordinates with a null map.');
+    assert(_googleMap!.projection != null,
+        'Cannot compute screen coordinate with a null map or projection.');
+
     final point =
-        _googleMap.projection.fromLatLngToPoint(_latLngToGmLatLng(latLng));
-    return ScreenCoordinate(x: point.x, y: point.y);
+        _googleMap!.projection!.fromLatLngToPoint!(_latLngToGmLatLng(latLng))!;
+
+    assert(point.x != null && point.y != null,
+        'The x and y of a ScreenCoordinate cannot be null.');
+
+    return ScreenCoordinate(x: point.x!.toInt(), y: point.y!.toInt());
   }
 
   /// Returns the [LatLng] for a `screenCoordinate` (in pixels) of the viewport.
   Future<LatLng> getLatLng(ScreenCoordinate screenCoordinate) async {
+    assert(_googleMap != null,
+        'Cannot get the lat, lng of a screen coordinate with a null map.');
+
     final gmaps.LatLng latLng =
-        _pixelToLatLng(_googleMap, screenCoordinate.x, screenCoordinate.y);
+        _pixelToLatLng(_googleMap!, screenCoordinate.x, screenCoordinate.y);
     return _gmLatLngToLatLng(latLng);
   }
 
   /// Applies a `cameraUpdate` to the current viewport.
   Future<void> moveCamera(CameraUpdate cameraUpdate) async {
-    return _applyCameraUpdate(_googleMap, cameraUpdate);
+    assert(_googleMap != null, 'Cannot update the camera of a null map.');
+
+    return _applyCameraUpdate(_googleMap!, cameraUpdate);
   }
 
   /// Returns the zoom level of the current viewport.
-  Future<double> getZoomLevel() async => _googleMap.zoom.toDouble();
+  Future<double> getZoomLevel() async {
+    assert(_googleMap != null, 'Cannot get zoom level of a null map.');
+    assert(_googleMap!.zoom != null,
+        'Zoom level should not be null. Is the map correctly initialized?');
+
+    return _googleMap!.zoom!.toDouble();
+  }
 
   // Geometry manipulation
 
   /// Applies [CircleUpdates] to the currently managed circles.
   void updateCircles(CircleUpdates updates) {
+    assert(
+        _circlesController != null, 'Cannot update circles after dispose().');
     _circlesController?.addCircles(updates.circlesToAdd);
     _circlesController?.changeCircles(updates.circlesToChange);
     _circlesController?.removeCircles(updates.circleIdsToRemove);
@@ -291,6 +340,8 @@ class GoogleMapController {
 
   /// Applies [PolygonUpdates] to the currently managed polygons.
   void updatePolygons(PolygonUpdates updates) {
+    assert(
+        _polygonsController != null, 'Cannot update polygons after dispose().');
     _polygonsController?.addPolygons(updates.polygonsToAdd);
     _polygonsController?.changePolygons(updates.polygonsToChange);
     _polygonsController?.removePolygons(updates.polygonIdsToRemove);
@@ -298,6 +349,8 @@ class GoogleMapController {
 
   /// Applies [PolylineUpdates] to the currently managed lines.
   void updatePolylines(PolylineUpdates updates) {
+    assert(_polylinesController != null,
+        'Cannot update polylines after dispose().');
     _polylinesController?.addPolylines(updates.polylinesToAdd);
     _polylinesController?.changePolylines(updates.polylinesToChange);
     _polylinesController?.removePolylines(updates.polylineIdsToRemove);
@@ -305,6 +358,8 @@ class GoogleMapController {
 
   /// Applies [MarkerUpdates] to the currently managed markers.
   void updateMarkers(MarkerUpdates updates) {
+    assert(
+        _markersController != null, 'Cannot update markers after dispose().');
     _markersController?.addMarkers(updates.markersToAdd);
     _markersController?.changeMarkers(updates.markersToChange);
     _markersController?.removeMarkers(updates.markerIdsToRemove);
@@ -312,22 +367,29 @@ class GoogleMapController {
 
   /// Shows the [InfoWindow] of the marker identified by its [MarkerId].
   void showInfoWindow(MarkerId markerId) {
+    assert(_markersController != null,
+        'Cannot show infowindow of marker [${markerId.value}] after dispose().');
     _markersController?.showMarkerInfoWindow(markerId);
   }
 
   /// Hides the [InfoWindow] of the marker identified by its [MarkerId].
   void hideInfoWindow(MarkerId markerId) {
+    assert(_markersController != null,
+        'Cannot hide infowindow of marker [${markerId.value}] after dispose().');
     _markersController?.hideMarkerInfoWindow(markerId);
   }
 
   /// Returns true if the [InfoWindow] of the marker identified by [MarkerId] is shown.
   bool isInfoWindowShown(MarkerId markerId) {
-    return _markersController?.isInfoWindowShown(markerId);
+    return _markersController?.isInfoWindowShown(markerId) ?? false;
   }
 
   // Cleanup
 
   /// Disposes of this controller and its resources.
+  ///
+  /// You won't be able to call many of the methods on this controller after
+  /// calling `dispose`!
   void dispose() {
     _widget = null;
     _googleMap = null;
