@@ -45,7 +45,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> updateMapOptions(
     Map<String, dynamic> optionsUpdate, {
-    @required int mapId,
+    required int mapId,
   }) async {
     _map(mapId).updateRawOptions(optionsUpdate);
   }
@@ -54,7 +54,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> updateMarkers(
     MarkerUpdates markerUpdates, {
-    @required int mapId,
+    required int mapId,
   }) async {
     _map(mapId).updateMarkers(markerUpdates);
   }
@@ -63,7 +63,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> updatePolygons(
     PolygonUpdates polygonUpdates, {
-    @required int mapId,
+    required int mapId,
   }) async {
     _map(mapId).updatePolygons(polygonUpdates);
   }
@@ -72,7 +72,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> updatePolylines(
     PolylineUpdates polylineUpdates, {
-    @required int mapId,
+    required int mapId,
   }) async {
     _map(mapId).updatePolylines(polylineUpdates);
   }
@@ -81,15 +81,15 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> updateCircles(
     CircleUpdates circleUpdates, {
-    @required int mapId,
+    required int mapId,
   }) async {
     _map(mapId).updateCircles(circleUpdates);
   }
 
   @override
   Future<void> updateTileOverlays({
-    @required Set<TileOverlay> newTileOverlays,
-    @required int mapId,
+    required Set<TileOverlay> newTileOverlays,
+    required int mapId,
   }) async {
     return; // Noop for now!
   }
@@ -97,7 +97,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> clearTileCache(
     TileOverlayId tileOverlayId, {
-    @required int mapId,
+    required int mapId,
   }) async {
     return; // Noop for now!
   }
@@ -106,7 +106,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> animateCamera(
     CameraUpdate cameraUpdate, {
-    @required int mapId,
+    required int mapId,
   }) async {
     return moveCamera(cameraUpdate, mapId: mapId);
   }
@@ -115,7 +115,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> moveCamera(
     CameraUpdate cameraUpdate, {
-    @required int mapId,
+    required int mapId,
   }) async {
     return _map(mapId).moveCamera(cameraUpdate);
   }
@@ -128,8 +128,8 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   /// pass full styles.
   @override
   Future<void> setMapStyle(
-    String mapStyle, {
-    @required int mapId,
+    String? mapStyle, {
+    required int mapId,
   }) async {
     _map(mapId).updateRawOptions({
       'styles': _mapStyles(mapStyle),
@@ -139,7 +139,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   /// Returns the bounds of the current viewport.
   @override
   Future<LatLngBounds> getVisibleRegion({
-    @required int mapId,
+    required int mapId,
   }) {
     return _map(mapId).getVisibleRegion();
   }
@@ -148,7 +148,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<ScreenCoordinate> getScreenCoordinate(
     LatLng latLng, {
-    @required int mapId,
+    required int mapId,
   }) {
     return _map(mapId).getScreenCoordinate(latLng);
   }
@@ -157,7 +157,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<LatLng> getLatLng(
     ScreenCoordinate screenCoordinate, {
-    @required int mapId,
+    required int mapId,
   }) {
     return _map(mapId).getLatLng(screenCoordinate);
   }
@@ -170,7 +170,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> showMarkerInfoWindow(
     MarkerId markerId, {
-    @required int mapId,
+    required int mapId,
   }) async {
     _map(mapId).showInfoWindow(markerId);
   }
@@ -183,7 +183,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<void> hideMarkerInfoWindow(
     MarkerId markerId, {
-    @required int mapId,
+    required int mapId,
   }) async {
     _map(mapId).hideInfoWindow(markerId);
   }
@@ -196,7 +196,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   @override
   Future<bool> isMarkerInfoWindowShown(
     MarkerId markerId, {
-    @required int mapId,
+    required int mapId,
   }) async {
     return _map(mapId).isInfoWindowShown(markerId);
   }
@@ -204,7 +204,7 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   /// Returns the zoom level of the `mapId`.
   @override
   Future<double> getZoomLevel({
-    @required int mapId,
+    required int mapId,
   }) {
     return _map(mapId).getZoomLevel();
   }
@@ -213,64 +213,64 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   // into the plugin
 
   @override
-  Stream<CameraMoveStartedEvent> onCameraMoveStarted({@required int mapId}) {
+  Stream<CameraMoveStartedEvent> onCameraMoveStarted({required int mapId}) {
     return _events(mapId).whereType<CameraMoveStartedEvent>();
   }
 
   @override
-  Stream<CameraMoveEvent> onCameraMove({@required int mapId}) {
+  Stream<CameraMoveEvent> onCameraMove({required int mapId}) {
     return _events(mapId).whereType<CameraMoveEvent>();
   }
 
   @override
-  Stream<CameraIdleEvent> onCameraIdle({@required int mapId}) {
+  Stream<CameraIdleEvent> onCameraIdle({required int mapId}) {
     return _events(mapId).whereType<CameraIdleEvent>();
   }
 
   @override
-  Stream<MarkerTapEvent> onMarkerTap({@required int mapId}) {
+  Stream<MarkerTapEvent> onMarkerTap({required int mapId}) {
     return _events(mapId).whereType<MarkerTapEvent>();
   }
 
   @override
-  Stream<InfoWindowTapEvent> onInfoWindowTap({@required int mapId}) {
+  Stream<InfoWindowTapEvent> onInfoWindowTap({required int mapId}) {
     return _events(mapId).whereType<InfoWindowTapEvent>();
   }
 
   @override
-  Stream<MarkerDragEndEvent> onMarkerDragEnd({@required int mapId}) {
+  Stream<MarkerDragEndEvent> onMarkerDragEnd({required int mapId}) {
     return _events(mapId).whereType<MarkerDragEndEvent>();
   }
 
   @override
-  Stream<PolylineTapEvent> onPolylineTap({@required int mapId}) {
+  Stream<PolylineTapEvent> onPolylineTap({required int mapId}) {
     return _events(mapId).whereType<PolylineTapEvent>();
   }
 
   @override
-  Stream<PolygonTapEvent> onPolygonTap({@required int mapId}) {
+  Stream<PolygonTapEvent> onPolygonTap({required int mapId}) {
     return _events(mapId).whereType<PolygonTapEvent>();
   }
 
   @override
-  Stream<CircleTapEvent> onCircleTap({@required int mapId}) {
+  Stream<CircleTapEvent> onCircleTap({required int mapId}) {
     return _events(mapId).whereType<CircleTapEvent>();
   }
 
   @override
-  Stream<MapTapEvent> onTap({@required int mapId}) {
+  Stream<MapTapEvent> onTap({required int mapId}) {
     return _events(mapId).whereType<MapTapEvent>();
   }
 
   @override
-  Stream<MapLongPressEvent> onLongPress({@required int mapId}) {
+  Stream<MapLongPressEvent> onLongPress({required int mapId}) {
     return _events(mapId).whereType<MapLongPressEvent>();
   }
 
   /// Disposes of the current map. It can't be used afterwards!
   @override
-  void dispose({@required int mapId}) {
-    _map(mapId)?.dispose();
+  void dispose({required int mapId}) {
+    _map(mapId).dispose();
     _mapById.remove(mapId);
   }
 
@@ -278,19 +278,16 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
   Widget buildView(
     int creationId,
     PlatformViewCreatedCallback onPlatformViewCreated, {
-    @required CameraPosition initialCameraPosition,
+    required CameraPosition initialCameraPosition,
     Set<Marker> markers = const <Marker>{},
     Set<Polygon> polygons = const <Polygon>{},
     Set<Polyline> polylines = const <Polyline>{},
     Set<Circle> circles = const <Circle>{},
     Set<TileOverlay> tileOverlays = const <TileOverlay>{},
-    Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers =
+    Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers =
         const <Factory<OneSequenceGestureRecognizer>>{},
     Map<String, dynamic> mapOptions = const <String, dynamic>{},
   }) {
-    assert(creationId != null,
-        'buildView needs a `_webOnlyMapCreationId` in its creationParams to prevent widget reloads in web.');
-
     // Bail fast if we've already rendered this map ID...
     if (_mapById[creationId]?.widget != null) {
       return _mapById[creationId].widget;
@@ -314,6 +311,9 @@ class GoogleMapsPlugin extends GoogleMapsFlutterPlatform {
 
     onPlatformViewCreated.call(creationId);
 
-    return mapController.widget;
+    assert(mapController.widget != null,
+        'The widget of a GoogleMapController cannot be null before calling dispose on it.');
+
+    return mapController.widget!;
   }
 }
