@@ -14,8 +14,8 @@ class MarkersController extends GeometryController {
 
   /// Initialize the cache. The [StreamController] comes from the [GoogleMapController], and is shared with other controllers.
   MarkersController({
-    @required StreamController<MapEvent> stream,
-  })  : _streamController = stream,
+    required StreamController<MapEvent> stream,
+  })   : _streamController = stream,
         _markerIdToController = Map<MarkerId, MarkerController>();
 
   /// Returns the cache of [MarkerController]s. Test only.
@@ -26,7 +26,7 @@ class MarkersController extends GeometryController {
   ///
   /// Wraps each [Marker] into its corresponding [MarkerController].
   void addMarkers(Set<Marker> markersToAdd) {
-    markersToAdd?.forEach(_addMarker);
+    markersToAdd.forEach(_addMarker);
   }
 
   void _addMarker(Marker marker) {
@@ -35,14 +35,15 @@ class MarkersController extends GeometryController {
     }
 
     final infoWindowOptions = _infoWindowOptionsFromMarker(marker);
-    gmaps.InfoWindow gmInfoWindow;
+    gmaps.InfoWindow? gmInfoWindow;
 
     if (infoWindowOptions != null) {
       gmInfoWindow = gmaps.InfoWindow(infoWindowOptions);
       // Google Maps' JS SDK does not have a click event on the InfoWindow, so
       // we make one...
       if (infoWindowOptions.content is HtmlElement) {
-        infoWindowOptions.content.onClick.listen((_) {
+        final content = infoWindowOptions.content as HtmlElement;
+        content.onClick.listen((_) {
           _onInfoWindowTap(marker.markerId);
         });
       }
@@ -70,11 +71,11 @@ class MarkersController extends GeometryController {
 
   /// Updates a set of [Marker] objects with new options.
   void changeMarkers(Set<Marker> markersToChange) {
-    markersToChange?.forEach(_changeMarker);
+    markersToChange.forEach(_changeMarker);
   }
 
   void _changeMarker(Marker marker) {
-    MarkerController markerController = _markerIdToController[marker?.markerId];
+    MarkerController? markerController = _markerIdToController[marker.markerId];
     if (markerController != null) {
       final markerOptions = _markerOptionsFromMarker(
         marker,
@@ -83,18 +84,18 @@ class MarkersController extends GeometryController {
       final infoWindow = _infoWindowOptionsFromMarker(marker);
       markerController.update(
         markerOptions,
-        newInfoWindowContent: infoWindow?.content,
+        newInfoWindowContent: infoWindow?.content as HtmlElement?,
       );
     }
   }
 
   /// Removes a set of [MarkerId]s from the cache.
   void removeMarkers(Set<MarkerId> markerIdsToRemove) {
-    markerIdsToRemove?.forEach(_removeMarker);
+    markerIdsToRemove.forEach(_removeMarker);
   }
 
   void _removeMarker(MarkerId markerId) {
-    final MarkerController markerController = _markerIdToController[markerId];
+    final MarkerController? markerController = _markerIdToController[markerId];
     markerController?.remove();
     _markerIdToController.remove(markerId);
   }
@@ -106,7 +107,7 @@ class MarkersController extends GeometryController {
   /// See also [hideMarkerInfoWindow] and [isInfoWindowShown].
   void showMarkerInfoWindow(MarkerId markerId) {
     _hideAllMarkerInfoWindow();
-    MarkerController markerController = _markerIdToController[markerId];
+    MarkerController? markerController = _markerIdToController[markerId];
     markerController?.showInfoWindow();
   }
 
@@ -114,7 +115,7 @@ class MarkersController extends GeometryController {
   ///
   /// See also [showMarkerInfoWindow] and [isInfoWindowShown].
   void hideMarkerInfoWindow(MarkerId markerId) {
-    MarkerController markerController = _markerIdToController[markerId];
+    MarkerController? markerController = _markerIdToController[markerId];
     markerController?.hideInfoWindow();
   }
 
@@ -122,7 +123,7 @@ class MarkersController extends GeometryController {
   ///
   /// See also [showMarkerInfoWindow] and [hideMarkerInfoWindow].
   bool isInfoWindowShown(MarkerId markerId) {
-    MarkerController markerController = _markerIdToController[markerId];
+    MarkerController? markerController = _markerIdToController[markerId];
     return markerController?.infoWindowShown ?? false;
   }
 
