@@ -6,15 +6,15 @@ part of google_maps_flutter_web;
 
 /// The `PolygonController` class wraps a [gmaps.Polygon] and its `onTap` behavior.
 class PolygonController {
-  gmaps.Polygon _polygon;
+  gmaps.Polygon? _polygon;
 
   final bool _consumeTapEvents;
 
   /// Creates a `PolygonController` that wraps a [gmaps.Polygon] object and its `onTap` behavior.
   PolygonController({
-    @required gmaps.Polygon polygon,
+    required gmaps.Polygon polygon,
     bool consumeTapEvents = false,
-    ui.VoidCallback onTap,
+    ui.VoidCallback? onTap,
   })  : _polygon = polygon,
         _consumeTapEvents = consumeTapEvents {
     if (onTap != null) {
@@ -26,20 +26,25 @@ class PolygonController {
 
   /// Returns the wrapped [gmaps.Polygon]. Only used for testing.
   @visibleForTesting
-  gmaps.Polygon get polygon => _polygon;
+  gmaps.Polygon? get polygon => _polygon;
 
   /// Returns `true` if this Controller will use its own `onTap` handler to consume events.
   bool get consumeTapEvents => _consumeTapEvents;
 
   /// Updates the options of the wrapped [gmaps.Polygon] object.
+  ///
+  /// This cannot be called after [remove].
   void update(gmaps.PolygonOptions options) {
-    _polygon.options = options;
+    assert(_polygon != null, 'Cannot `update` Polygon after calling `remove`.');
+    _polygon!.options = options;
   }
 
   /// Disposes of the currently wrapped [gmaps.Polygon].
   void remove() {
-    _polygon.visible = false;
-    _polygon.map = null;
-    _polygon = null;
+    if (_polygon != null) {
+      _polygon!.visible = false;
+      _polygon!.map = null;
+      _polygon = null;
+    }
   }
 }
